+++ report.go
@@ -19,9 +19,9 @@ func displayReport(results []Result) {
 		fmt.Printf("Status:  %d (%s)\n", r.StatusCode, http.StatusText(r.StatusCode))
 		if len(referrers[r.URL]) > 0 {
 			if len(referrers[r.URL]) > 3 {
-				fmt.Printf("Refs: %s ... (%dx)\n", strings.Join(referrers[r.URL][0:3], "\n           "), len(referrers[r.URL]))
+				fmt.Printf("Refs:    %s ... (%dx)\n", strings.Join(referrers[r.URL][0:3], "\n         "), len(referrers[r.URL]))
 			} else {
-				fmt.Printf("Refs: %s\n", strings.Join(referrers[r.URL], "\n           "))
+				fmt.Printf("Refs:    %s\n", strings.Join(referrers[r.URL], "\n         "))
 			}
 		}