+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [dev]
+
+- Fix report referer formatting
+
 ## [0.0.2]
 
 - Improve error handling